+++ bindings/c/include/manifoldc.h
@@ -1,6 +1,7 @@
 #include <stddef.h>
 #include <stdint.h>
-#include <types.h>
+
+#include "types.h"
 
 #ifdef __cplusplus
 #include <cstdint>
