+++ bindings/c/manifoldc.cpp
@@ -29,8 +29,7 @@ ManifoldMeshGL *level_set(void *mem, float (*sdf)(float, float, float),
   std::function<float(glm::vec3)> fun = [sdf](glm::vec3 v) {
     return (sdf(v.x, v.y, v.z));
   };
-  auto pol = seq ? std::make_optional(ExecutionPolicy::Seq) : std::nullopt;
-  auto mesh = LevelSet(fun, *from_c(bounds), edge_length, level, pol);
+  auto mesh = LevelSet(fun, *from_c(bounds), edge_length, level, !seq);
   return to_c(new (mem) MeshGL(mesh));
 }
 ManifoldMeshGL *level_set_context(
@@ -43,8 +42,7 @@ ManifoldMeshGL *level_set_context(
   std::function<float(glm::vec3)> fun = [sdf](glm::vec3 v) {
     return (sdf(v.x, v.y, v.z));
   };
-  auto pol = seq ? std::make_optional(ExecutionPolicy::Seq) : std::nullopt;
-  auto mesh = LevelSet(fun, *from_c(bounds), edge_length, level, pol);
+  auto mesh = LevelSet(fun, *from_c(bounds), edge_length, level, !seq);
   return to_c(new (mem) MeshGL(mesh));
 }
 }  // namespace
