+++ bindings/c/meshio.cpp
@@ -1,8 +1,7 @@
 #include "meshIO.h"
 
-#include <conv.h>
-
-#include "include/types.h"
+#include "conv.h"
+#include "types.h"
 
 // C <-> C++ conversions
