+++ bindings/c/rect.cpp
@@ -1,8 +1,7 @@
-#include <conv.h>
-#include <manifold.h>
-#include <public.h>
-
+#include "conv.h"
 #include "cross_section.h"
+#include "manifold.h"
+#include "public.h"
 #include "types.h"
 
 ManifoldRect *manifold_rect(void *mem, float x1, float y1, float x2, float y2) {
