+++ bindings/wasm/bindings.cpp
@@ -14,14 +14,14 @@
 
 #include <emscripten/bind.h>
 #include <emscripten/val.h>
-#include <manifold.h>
-#include <polygon.h>
-#include <sdf.h>
 
 #include <vector>
 
 #include "cross_section.h"
 #include "helpers.cpp"
+#include "manifold.h"
+#include "polygon.h"
+#include "sdf.h"
 
 using namespace emscripten;
 using namespace manifold;
