+++ test/test_main.cpp
@@ -126,7 +126,7 @@ Mesh Csaszar() {
 }
 
 struct GyroidSDF {
-  __host__ __device__ float operator()(glm::vec3 p) const {
+  float operator()(glm::vec3 p) const {
     const glm::vec3 min = p;
     const glm::vec3 max = glm::vec3(glm::two_pi<float>()) - p;
     const float min3 = glm::min(min.x, glm::min(min.y, min.z));
